% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Event.R
\name{add.identifier}
\alias{add.identifier}
\title{Add an identifier to an \code{instrument}}
\usage{
add.identifier(primary_id, ...)
}
\arguments{
\item{primary_id}{primary_id of an \code{\link{instrument}}}

\item{...}{identifiers passed as regular named arguments.}
}
\value{
called for side-effect
}
\description{
Add an identifier to an \code{\link{instrument}} unless the instrument
already has that identifier.
}
\examples{
\dontrun{
stock("XXX", currency("USD"))
add.identifier("XXX", yahoo="^XXX")
getEvent("^XXX")
add.identifier("^XXX", "x3")
all.equal(getEvent("x3"), getEvent("XXX")) #TRUE
}
}
\author{
Garrett See
}
\seealso{
\code{\link{event_attr}}
}

