% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Event.R
\name{addEventDefinedBy}
\alias{addEventDefinedBy}
\title{Add a source to the defined.by field of an \code{event}}
\usage{
addEventDefinedBy(primary_ids, ...)
}
\arguments{
\item{primary_ids}{character vector of primary_ids of
\code{\link{event}}s}

\item{...}{strings, or character vector, or semi-colon delimited string.}
}
\value{
called for side-effect
}
\description{
Concatenate a string or strings (passed through dots) to the defined.by
field of an event (separated by semi-colons).  Any duplicates will be
removed.  See Details.
}
\details{
If there is already a value for the \code{defined.by} attribute of the
\code{primary_id} event, that string will be split on semi-colons and
converted to a character vector.  That will be \code{c}ombined with any new
strings (in \code{...}).  The unique value of this new vector will then
be converted into a semi-colon delimited string that will be assigned to
the \code{defined.by} attribute of the \code{primary_ids}' events

Many functions that create or update event definitions will also add or
update the value of the defined.by attribute of that event.  If an
event has been updated by more than one function, it's \code{defined.by}
attribute will likely be a semi-colon delimited string (e.g.
\dQuote{TTR;yahoo}).
}
\examples{
\dontrun{
update_events.TTR("GS")
getEvent("GS")$defined.by #TTR
addEventDefinedBy("GS", "gsee", "demo")
addEventDefinedBy("GS", "gsee;demo") #same
}
}
\author{
Gei Lin
}
\seealso{
\code{\link{addEventIdentifier}}, \code{\link{event_attr}}
}

