% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Event.R
\name{addEventIdentifier}
\alias{addEventIdentifier}
\title{Add an event identifier to an \code{event}}
\usage{
addEventIdentifier(primary_id, ...)
}
\arguments{
\item{primary_id}{primary_id of an \code{\link{event}}}

\item{...}{identifiers passed as regular named arguments.}
}
\value{
called for side-effect
}
\description{
Add an identifier to an \code{\link{event}} unless the event
already has that identifier.
}
\examples{
\dontrun{
auction("XXX", region("USD"))
add.identifier("XXX", yahoo="^XXX")
getEvent("^XXX")
add.identifier("^XXX", "x3")
all.equal(getEvent("x3"), getEvent("XXX")) #TRUE
}
}
\author{
Garrett See
}
\seealso{
\code{\link{event_attr}}
}

