% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EventInterval.R
\name{EventInterval}
\alias{Break}
\alias{Concession}
\alias{EventInterval}
\alias{Settle2Settle}
\title{synthetic event constructors}
\usage{
EventInterval(primary_id, start, end, anchor, startn = 1, endn = 1,
  startoff = 0, endoff = 0, ..., identifiers = NULL, assign_i = TRUE,
  type = c("eventinterval"))

Concession(primary_id, auction = NULL, settle = NULL, identifiers = NULL,
  assign_i = TRUE, overwrite = TRUE, ...)

Break(primary_id, auction = NULL, settle = NULL, identifiers = NULL,
  assign_i = TRUE, overwrite = TRUE, ...)

Settle2Settle(primary_id, settle = NULL, identifiers = NULL,
  assign_i = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{primary_id}{chr string of primary identifier of event to be defined.}

\item{start}{starting event}

\item{end}{ending event}

\item{anchor}{anchor event that dictates the starting event}

\item{startn}{instance offset for the starting event}

\item{endn}{instance offset for the ending event}

\item{startoff}{offset from the starting event in seconds, positive or negative}

\item{endoff}{offset from the ending event in seconds, positive or negative}

\item{identifiers}{identifiers}

\item{assign_i}{TRUE/FALSE. Should the event be assigned in the \code{.event} environment?}

\item{type}{of eventinterval; wrappers do not require this.}

\item{overwrite}{if FALSE and an event with the same \code{primary_id}
is already defined, an error will be thrown and no events will be
created.}

\item{region}{chr string name of region denomination}

\item{\dots}{any other passthrough parameters}
}
\value{
called for side effect. stores an event in .event environment
}
\description{
define EventInterval, and other synthetic events wrapped around EventInterval
}
\details{
EventIntervals have one underlying or anchor event.

The \code{suffix_id} parameter of wrapper functions such as  \code{guaranteed_spread} is presumed to
be a string describing the \code{members}.
It will be \code{\link{strsplit}} using the regex "[-;:_,\\.]" to create the \code{members} vector,
and potentially combined with a \code{root_id}.

Most wrappers will build \code{primary_id} if it is NULL, either by combining \code{root_id} and \code{suffix_id}, or
by passing \code{members} in a call to \code{\link{make_spread_id}}
}
\examples{
\dontrun{
Auction('US.10YR.FIX','US')
OpenClose('US.CBTN.REG.CLOSE.POST','US')
EventInterval('US.10YR.FIX','US.CBTN.REG.CLOSE.POST','US')
}
}
\author{
Gei Lin
}
\seealso{
Event, Auction, Policy, EcoData
}

