% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Exchange.R
\name{getExchange}
\alias{getExchange}
\title{Primary accessor function for getting objects of class 'exchange'}
\usage{
getExchange(x, silent = FALSE, type = "exchange")
}
\arguments{
\item{x}{String identifier of exchange to retrieve}

\item{silent}{if TRUE, will not warn on failure, default FALSE}

\item{type}{class of object to look for. See Details}
}
\description{
This function will search the \code{.exchange} environment for objects of
class \code{type}, using first the \code{primary_id} and then any
\code{identifiers} to locate the exchange.
}
\details{
\code{getRegion} will try using each possible \code{primary_id}
until it finds an exchange of the appropriate \code{type}
}
\examples{
\dontrun{
Exchange('CBOT', region = 'US')

getExchange("CBOT", Region("US"))
}
}

