% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Region.R
\name{getRegion}
\alias{getRegion}
\title{Primary accessor function for getting objects of class 'region'}
\usage{
getRegion(x, silent = FALSE, type = "region")
}
\arguments{
\item{x}{String identifier of region to retrieve}

\item{silent}{if TRUE, will not warn on failure, default FALSE}

\item{type}{class of object to look for. See Details}
}
\description{
This function will search the \code{.region} environment for objects of
class \code{type}, using first the \code{primary_id} and then any
\code{identifiers} to locate the region.
}
\details{
\code{getRegion} will try using each possible \code{primary_id}
until it finds an region of the appropriate \code{type}
}
\examples{
\dontrun{
Region('US')

getRegion("US")
getRegion('EU')
}
}

