% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Event.R
\name{is.event}
\alias{is.event}
\title{class test for object supposedly of type 'event'}
\usage{
is.event(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'event'
}

