% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Exchange.R
\name{is.exchange.name}
\alias{is.exchange.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{Exchange}}}
\usage{
is.exchange.name(x)
}
\arguments{
\item{x}{character vector}
}
\description{
check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{Exchange}}
}

