% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Exchange.R
\name{is.exchange}
\alias{is.exchange}
\title{class test for object supposedly of type 'exchange'}
\usage{
is.exchange(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'exchange'
}

