% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Market.R
\name{is.market}
\alias{is.market}
\title{class test for object supposedly of type 'market'}
\usage{
is.market(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'market'
}

