% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Market.R
\name{is.region.name}
\alias{is.region.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{exchange}}}
\usage{
is.region.name(x)
}
\arguments{
\item{x}{character vector}
}
\description{
check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{exchange}}
}

