% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Region.R
\name{is.region}
\alias{is.region}
\title{class test for object supposedly of type 'region'}
\usage{
is.region(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'region'
}

