% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Market.R
\name{is.session}
\alias{is.session}
\title{class test for object supposedly of type 'session'}
\usage{
is.session(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'session'
}

