% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ls_exchanges.R
\name{ls_exchanges}
\alias{ls_exchanges}
\alias{rm_exchanges}
\title{List or Remove exchange objects}
\usage{
ls_exchanges(pattern = NULL, match = TRUE, verbose = TRUE)
}
\arguments{
\item{pattern}{an optional regular expression.  Only names matching
\sQuote{pattern} are returned.}

\item{match}{return only exact matches?}

\item{verbose}{be verbose?}

\item{x}{what to remove. if not supplied all exchanges of relevent class
will be removed.  For \code{ls_defined.by} x is the string describing how the
exchange was defined.}
}
\value{
ls functions return vector of character strings corresponding to
exchanges of requested type rm functions are called for side-effect
}
\description{
display the names of or delete exchanges, auctions, options, futures,
exchanges, bonds, funds, spreads, guaranteed_spreads, synthetics,
derivatives, or non-derivatives.
}
\details{
ls functions return the names of all the exchanges of the class implied by
the function name. rm functions remove the exchanges of the class implied
by the function name

rm_exchanges and rm_non_derivatives will not delete exchanges unless the
keep.exchanges argument is FALSE.

For the rm functions, x can be a vector of exchange names, or nothing.  If
\code{x} is missing, all exchanges of the relevant type will be removed.

It can be useful to nest these functions to get things like futures
denominated in USD.
}
\examples{
\dontrun{

# First, create some exchanges
Exchange(c("NYSE", "COMEX", "ICE", "CBOT"))

# Now, the examples
ls_exchanges() #all exchanges
rm_exchanges()
}
}
\author{
Gei Lin
}
\seealso{
ls, rm, Exchange
}

