% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ls_regions.R
\name{ls_regions}
\alias{ls_regions}
\alias{rm_regions}
\title{List or Remove region objects}
\usage{
ls_regions(pattern = NULL, match = TRUE, verbose = TRUE)
}
\arguments{
\item{pattern}{an optional regular expression.  Only names matching
\sQuote{pattern} are returned.}

\item{match}{return only exact matches?}

\item{verbose}{be verbose?}

\item{x}{what to remove. if not supplied all regions of relevent class
will be removed.  For \code{ls_defined.by} x is the string describing how the
region was defined.}
}
\value{
ls functions return vector of character strings corresponding to
regions of requested type rm functions are called for side-effect
}
\description{
display the names of or delete regions, auctions, options, futures,
regions, bonds, funds, spreads, guaranteed_spreads, synthetics,
derivatives, or non-derivatives.
}
\details{
ls functions return the names of all the regions of the class implied by
the function name. rm functions remove the regions of the class implied
by the function name

rm_regions and rm_non_derivatives will not delete regions unless the
keep.regions argument is FALSE.

For the rm functions, x can be a vector of region names, or nothing.  If
\code{x} is missing, all regions of the relevant type will be removed.

It can be useful to nest these functions to get things like futures
denominated in USD.
}
\examples{
\dontrun{

# First, create some regions
Region(c("US", "EU", "JP", "FR"))

# Now, the examples
ls_regions() #all regions
rm_regions()
}
}
\author{
Gei Lin
}
\seealso{
ls, rm, Region
}

