% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Market.R
\name{Market}
\alias{Broken}
\alias{Continuous}
\alias{Exchange}
\alias{Market}
\alias{Region}
\alias{Session}
\title{market class constructors}
\usage{
Market(primary_id, ..., exchange, identifiers = NULL, type = NULL,
  assign_i = FALSE, overwrite = TRUE)

Continuous(primary_id, exchange, identifiers = NULL, assign_i = TRUE,
  overwrite = TRUE, ...)

Broken(primary_id, exchange, identifiers = NULL, assign_i = TRUE,
  overwrite = TRUE, ...)
}
\arguments{
\item{primary_id}{String describing the unique ID for the market. Most
of the wrappers allow this to be a vector.}

\item{...}{Any other passthru parameters, including}

\item{exchange}{String describing the exchange ID of an object of type
\code{\link{exchange}}}

\item{identifiers}{Named list of any other identifiers that should also be
stored for this market}

\item{type}{market type to be appended to the class definition, typically
"continuous" or "broken"}

\item{assign_i}{TRUE/FALSE. Should the market be assigned to the
\code{.market} environment?  Default is FALSE for \code{market},
TRUE for wrappers.}

\item{overwrite}{TRUE/FALSE. Should existing markets with the same
primary_id be overwritten? Default is TRUE. If FALSE, an error will be
thrown and the market will not be created.}
}
\description{
All 'exchange' and 'session' events must be defined before events of other types
may be defined.
}
\details{
Regions must also be defined -- all events have a region attached including sessions
E.g. US.CBTN is a based on the US region

In \dots you may pass any other arbitrary market fields that will be used
to create 'custom' fields.  S3 classes in \R are basically lists with a class
attribute.  We use this to our advantage to allow us to set arbitrary fields.

\code{identifiers} should be a named list to specify other identifiers beyond
the \code{primary_id}.  Please note that whenever possible, these should
still be unique.  Perhaps Bloomberg, Reuters-X.RIC, CUSIP, etc.
\code{\link{getEvent}} will return the first (and only the first) match
that it finds, starting with the primary_id, and then searching the
primary_ids of all markets for each of the \code{identifiers}.  Note that
when a large number of markets are defined, it is faster to find
markets by \code{primary_id} than by \code{identifiers} because it looks
for \code{primary_id}s first.

The \code{primary_id} will be coerced within reason to a valid \R variable
name by using \code{\link{make.names}}. We also remove any leading '1' digit
(a simple workaround to account for issues with the Reuters API).  If you are
defining an market that is not a \code{session}, with a primary_id that
already belongs to a \code{session}, a new primary_id will be create using
\code{make.names}.  For example, \code{auction("USD", Region("USD"))}, would
create a auction with a primary_id of \dQuote{USD.1} instead of overwritting
the \code{session}.

Please use some care to choose your primary identifiers so that R won't
complain.  If you have better regular expression code, we'd be happy to
include it.

\code{assign_i} will use \code{\link{assign}} to place the constructed
market class object into the \code{.market} environment.  Most of the
special type-specific constructors will use \code{assign_i=TRUE} internally.
Calling with \code{assign_i=FALSE}, or not specifying it, will return an
object and will \emph{not} store it.  Use this option ether to wrap calls to
\code{market} prior to further processing (and presumably assignment) or
to test your parameters before assignment.

If \code{overwrite=FALSE} is used, an error will be thrown if any
\code{primary_id}s are already in use.

As of version 0.10.0, the .market environment is located at the top level
of the package. i.e. \code{.market}.

\code{Continuous} and \code{Broken} are used to define the contract specs of a
series of markets.  The \code{primary_id} for these can begin with 1 or
2 dots if you need to avoid overwriting another market.
For example, if you have a \code{session} with \sQuote{CBTN} as the
\code{primary_id}, you could use \sQuote{.CBTN} as the \code{primary_id} of
the regular session \code{REG} specs, and \sQuote{..CBTN} as the \code{primary_id} of the
electronic session \code{ALL} specs. (or vice versa)

You can (optionally) provide a \code{src} argument in which case, it will be
used in a call to \code{\link[quantmod]{setSymbolLookup}}.
}
\seealso{
\code{\link{Continuous}},
\code{\link{Broken}},
\code{\link{MarketSession}},
\code{\link{load.markets}}
}

