% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\title{region class constructors}
\usage{
Region(primary_id, assign_i = TRUE, overwrite = TRUE)
}
\arguments{
\item{primary_id}{String describing the unique ID for the region. Most
of the wrappers allow this to be a vector.}

\item{assign_i}{TRUE/FALSE. Should the region be assigned to the
\code{.region} environment?  Default is TRUE for \code{Region}, and
TRUE for wrappers.}

\item{overwrite}{TRUE/FALSE. Should existing regions with the same
primary_id be overwritten? Default is TRUE. If FALSE, an error will be
thrown and the region will not be created.}
}
\description{
All regions must be defined before markets can be created
}
\details{
All exchanges have a region attached including sessions
E.g. CBOT is a based in the US region

\code{\link{getRegion}} will return the first (and only the first) match
that it finds, starting with the primary_id, and then searching the
primary_ids of all exchanges for each of the \code{identifiers}.  Note that
when a large number of exchanges are defined, it is faster to find
exchanges by \code{primary_id} than by \code{identifiers} because it looks
for \code{primary_id}s first.

The \code{primary_id} will be coerced within reason to a valid \R variable
name by using \code{\link{make.names}}.

Please use some care to choose your primary identifiers so that R won't
complain.  If you have better regular expression code, we'd be happy to
include it.

\code{assign_i} will use \code{\link{assign}} to place the constructed
region class object into the \code{.region} environment.  Most of the
special type-specific constructors will use \code{assign_i=TRUE} internally.
Calling with \code{assign_i=FALSE}, or not specifying it, will return an
object and will \emph{not} store it.  Use this option ether to wrap calls to
\code{region} prior to further processing (and presumably assignment) or
to test your parameters before assignment.

If \code{overwrite=FALSE} is used, an error will be thrown if any
\code{primary_id}s are already in use.
}
\seealso{
\code{\link{Region}},
\code{\link{load.regions}}
}

