% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Event.R
\name{sort.event}
\alias{sort.event}
\title{Event class sort method}
\usage{
\method{sort}{event}(x, decreasing = FALSE, na.last = NA, ...)
}
\description{
Event class sort method
}
\author{
Gei Lin
}
\keyword{internal}

