% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EventInterval.R
\name{EventInterval}
\alias{EventInterval}
\alias{butterfly}
\alias{guaranteed_spread}
\alias{spread}
\alias{synthetic}
\alias{synthetic.event}
\title{synthetic event constructors}
\usage{
EventInterval(primary_id, region, start, end, anchor, startn = 1, endn = 1,
  startoff = 0, endoff = 0, ..., multiplier = 1, tick_size = NULL,
  identifiers = NULL, assign_i = TRUE, type = c("eventinterval"))
}
\arguments{
\item{primary_id}{chr string of primary identifier of event to be defined.}

\item{region}{chr string name of region denomination}

\item{start}{}

\item{end}{}

\item{anchor}{}

\item{startn}{}

\item{endn}{}

\item{startoff}{}

\item{endoff}{}

\item{multiplier}{multiplier of the spread (1 / divisor for price weighted baskets)}

\item{tick_size}{minimum price change of the spread}

\item{identifiers}{identifiers}

\item{assign_i}{TRUE/FALSE. Should the event be assigned in the \code{.event} environment?}

\item{type}{type of event; wrappers do not require this.}

\item{\dots}{any other passthrough parameters}

\item{overwrite}{if FALSE and an event with the same \code{primary_id}
is already defined, an error will be thrown and no events will be
created.}
}
\value{
called for side effect. stores an event in .event environment
}
\description{
define EventInterval, and other synthetic events
}
\details{
EventIntervals have one underlying or anchor event.

The \code{suffix_id} parameter of wrapper functions such as  \code{guaranteed_spread} is presumed to
be a string describing the \code{members}.
It will be \code{\link{strsplit}} using the regex "[-;:_,\\.]" to create the \code{members} vector,
and potentially combined with a \code{root_id}.

Most wrappers will build \code{primary_id} if it is NULL, either by combining \code{root_id} and \code{suffix_id}, or
by passing \code{members} in a call to \code{\link{make_spread_id}}

\code{ICS} will build an Intercommodity Spread.  Although the expiration date and ratio may change,
the members of a given ICS will not change.  Therefore, \code{ICS_root} can be used to hold the
members of an Intercommodity Spread.  If an \code{ICS_root} has not been defined, then \code{members}
will be a required argument for \code{ICS}

We welcome assistance from others to model more complex OTC derivatives such as swap products.
}
\examples{
\dontrun{
Auction('US.10YR.FIX','US')
OpenClose('US.CBTN.REG.CLOSE.POST','US')
EventInterval('US.10YR.FIX','US.CBTN.REG.CLOSE.POST','US')
}
}
\author{
Gei Lin
}
\seealso{
Event, Auction
}

