% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Event.R
\name{is.event.name}
\alias{is.event.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of an \code{\link{event}}}
\usage{
is.event.name(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
check each element of a character vector to see if it is either the
primary_id or an identifier of an \code{\link{event}}
}

